Inline loop
The number of ports is fixed for a given scan
operation, so you don't need a list. Since
keeping a sequence is convenient we can stick to
the tuple (instead of a set, which would have
been even better if keeping the sequence was of
no consequence). Further, the loop can be
simplified.