Optimize CLIArgumentsParser.parse_ports
It's important to limit an implementation to its
most basic function. A "parse_ports" method
simply parses a string of comma-separated ports,
nothing more. It doesn't need to do the parsing
AND change the object's state by handling the
value of self.ports. Let's limit the method to
doing the parsing and then assign its returned
value to the instance attribute in another step.
Additionally, let's implement it as an Iterator
so that it doesn't even bear the responsibility
of determining what kind of data structure will
be used in the end.

You can see that we decide in the "parse" method
that the adequate data structure is a tuple, but
not before it. The "parse_ports" method remains
pure, so to speak. It has nothing to do with how
the data will be represented afterwards.