Fixed bug
The socket was not being closed since the last refactoring. This
was making it so that it would only connect with the first open
port it found, and not discover any further open ports. I moved
the context manager inside the for loop that scans the ports so
that the socket is closed every time it scans a port.
The ports are now being scanned correctly again and the program
shows the right output.